fixup adjust imports after splitting models
This is a technical commit to allow for easier review,
should be squashed after review.
