+++ tests/conftest.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 import pytest
 
-from unblob.models import Task, TaskResult
+from unblob.tasks import Task, TaskResult
 from unblob.testing import (  # noqa: F401 (module imported but unused)
     configure_logging,
     extraction_config,
